requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
ImageIO==2.37.3
kiwisolver==1.5.0
lazy-loader==0.5
matplotlib==3.10.9
networkx==3.6.1
numpy==2.4.4
packaging==26.2
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-image==0.26.0
scipy==1.17.1
six==1.17.0
tifffile==2026.5.2
